optimization changes, made it so it was all doubles instead of converting from ints to doubles, additionally fixed an issue where an invalid input would cause an infinite loop

+++ main.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 extern "C" {
     double circle();
-    int isfloat();
+    double isfloat();
 }
 
 int main() {
